fix calloc

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: tmuranak <tmuranak@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/20 18:06:45 by tmuranak          #+#    #+#             */
-/*   Updated: 2024/04/27 20:45:49 by tmuranak         ###   ########.fr       */
+/*   Updated: 2024/04/30 18:53:38 by tmuranak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 void	*ft_calloc(size_t count, size_t size)
 {
-	void	*rbuf;
-	int		buf_size;
+	void		*rbuf;
+	size_t		buf_size;
 
 	if (size > 0 && count > SIZE_MAX / size)
 		return (NULL);
